test: some cleanup re: #12890
