+++ lib/cursor/ChangeStream.js
@@ -54,6 +54,12 @@ class ChangeStream extends EventEmitter {
 
         ['close', 'change', 'end', 'error'].forEach(ev => {
           this.driverChangeStream.on(ev, data => {
+            // Sometimes Node driver still polls after close, so
+            // avoid any uncaught exceptions due to closed change streams
+            // See tests for gh-7022
+            if (ev === 'error' && this.closed) {
+              return;
+            }
             if (data != null && data.fullDocument != null && this.options && this.options.hydrate) {
               data.fullDocument = this.options.model.hydrate(data.fullDocument);
             }
@@ -71,6 +77,12 @@ class ChangeStream extends EventEmitter {
 
     ['close', 'change', 'end', 'error'].forEach(ev => {
       this.driverChangeStream.on(ev, data => {
+        // Sometimes Node driver still polls after close, so
+        // avoid any uncaught exceptions due to closed change streams
+        // See tests for gh-7022
+        if (ev === 'error' && this.closed) {
+          return;
+        }
         this.emit(ev, data);
       });
     });
