+++ test/collection.capped.test.js
@@ -19,15 +19,12 @@ const Schema = mongoose.Schema;
 describe('collections: capped:', function() {
   let db;
 
-  const connectionsToClose = [];
-
-  before(function() {
-    db = start();
-  });
-
-  after(async function() {
+  afterEach(async function() {
+    if (db == null) {
+      return;
+    }
     await db.close();
-    await Promise.all(connectionsToClose.map((v) => v.close()));
+    db = null;
   });
 
   it('schemas should have option size', function() {
@@ -40,6 +37,8 @@ describe('collections: capped:', function() {
 
   it('creation', async function() {
     this.timeout(15000);
+
+    db = start();
 
     await db.dropCollection('Test').catch(() => {});
 
@@ -55,7 +54,6 @@ describe('collections: capped:', function() {
 
   it('skips when setting autoCreate to false (gh-8566)', async function() {
     const db = start();
-    connectionsToClose.push(db);
     this.timeout(30000);
     await db.dropDatabase();
