+++ test/query.middleware.test.js
@@ -14,15 +14,12 @@ describe('query middleware', function() {
   let Author;
   let Publisher;
 
-  const connectionsToClose = [];
-
   before(function() {
     db = start();
   });
 
   after(async function() {
     await db.close();
-    await Promise.all(connectionsToClose.map((v) => v.close()));
   });
 
   const initializeData = function(done) {
@@ -83,9 +80,6 @@ describe('query middleware', function() {
       ++count;
       next();
     });
-
-    const conn = start();
-    connectionsToClose.push(conn);
 
     initializeData(function(error) {
       assert.ifError(error);
